NDRS-833: Move `SignatureCache` to a file.
